Move safepoints and stack maps code to its own module
This is purely mechanical code motion.
